+++ crates/rust-client/src/rpc/mod.rs
@@ -134,7 +134,7 @@ pub trait NodeRpcClient {
 
     /// Fetches note-related data for a list of [NoteId] using the `/GetNotesById` rpc endpoint
     ///
-    /// For any NoteType::Offchain note, the return data is only the [NoteMetadata], whereas
+    /// For any NoteType::Private note, the return data is only the [NoteMetadata], whereas
     /// for NoteType::Onchain notes, the return data includes all details.
     async fn get_notes_by_id(&mut self, note_ids: &[NoteId]) -> Result<Vec<NoteDetails>, RpcError>;