+++ tests/integration/custom_transactions_tests.rs
@@ -1,8 +1,6 @@
-use std::collections::BTreeMap;
-
 use miden_client::{
     accounts::AccountTemplate, transactions::transaction_request::TransactionRequest,
-    utils::Serializable,
+    utils::Serializable, ZERO,
 };
 use miden_objects::{
     accounts::{AccountId, AccountStorageType, AuthSecretKey},
@@ -10,6 +8,7 @@ use miden_objects::{
     assets::{FungibleAsset, TokenSymbol},
     crypto::{
         hash::rpo::Rpo256,
+        merkle::{MerkleStore, MerkleTree, NodeIndex},
         rand::{FeltRng, RpoRandomCoin},
     },
     notes::{
@@ -81,7 +80,7 @@ async fn test_transaction_request() {
     // these exact arguments
     let note_args_commitment = Rpo256::hash_elements(&NOTE_ARGS);
 
-    let note_args_map = BTreeMap::from([(note.id(), Some(note_args_commitment.into()))]);
+    let note_args_map = vec![(note.id(), Some(note_args_commitment.into()))];
     let mut advice_map = AdviceMap::new();
     advice_map.insert(note_args_commitment, NOTE_ARGS.to_vec());
 
@@ -117,16 +116,10 @@ async fn test_transaction_request() {
         client.compile_tx_script(program, script_inputs, vec![]).unwrap()
     };
 
-    let transaction_request = TransactionRequest::new(
-        regular_account.id(),
-        vec![],
-        note_args_map.clone(),
-        vec![],
-        vec![],
-        Some(tx_script),
-        Some(advice_map.clone()),
-    )
-    .unwrap();
+    let transaction_request = TransactionRequest::new(regular_account.id())
+        .with_authenticated_input_notes(note_args_map.clone())
+        .with_tx_script(tx_script)
+        .extend_advice_map(advice_map.clone());
 
     // This fails becuase of {asserted_value} having the incorrect number passed in
     assert!(client.new_transaction(transaction_request).is_err());
@@ -149,16 +142,116 @@ async fn test_transaction_request() {
         client.compile_tx_script(program, script_inputs, vec![]).unwrap()
     };
 
-    let transaction_request = TransactionRequest::new(
-        regular_account.id(),
-        vec![],
-        note_args_map,
-        vec![],
-        vec![],
-        Some(tx_script),
-        Some(advice_map),
-    )
-    .unwrap();
+    let transaction_request = TransactionRequest::new(regular_account.id())
+        .with_authenticated_input_notes(note_args_map)
+        .with_tx_script(tx_script)
+        .extend_advice_map(advice_map);
+
+    execute_tx_and_sync(&mut client, transaction_request).await;
+
+    client.sync_state().await.unwrap();
+}
+
+#[tokio::test]
+async fn test_merkle_store() {
+    let mut client = create_test_client();
+    wait_for_node(&mut client).await;
+
+    let account_template = AccountTemplate::BasicWallet {
+        mutable_code: false,
+        storage_type: AccountStorageType::OffChain,
+    };
+
+    client.sync_state().await.unwrap();
+    // Insert Account
+    let (regular_account, _seed) = client.new_account(account_template).unwrap();
+
+    let account_template = AccountTemplate::FungibleFaucet {
+        token_symbol: TokenSymbol::new("TEST").unwrap(),
+        decimals: 5u8,
+        max_supply: 10_000u64,
+        storage_type: AccountStorageType::OffChain,
+    };
+    let (fungible_faucet, _seed) = client.new_account(account_template).unwrap();
+
+    // Execute mint transaction in order to increase nonce
+    let note = mint_custom_note(&mut client, fungible_faucet.id(), regular_account.id()).await;
+    client.sync_state().await.unwrap();
+
+    // Prepare custom merkle store transaction
+
+    // If these args were to be modified, the transaction would fail because the note code expects
+    // these exact arguments
+    let note_args_commitment = Rpo256::hash_elements(&NOTE_ARGS);
+
+    let note_args_map = vec![(note.id(), Some(note_args_commitment.into()))];
+    let mut advice_map = AdviceMap::new();
+    advice_map.insert(note_args_commitment, NOTE_ARGS.to_vec());
+
+    // Build merkle store and advice stack with merkle root
+    let leaves: Vec<Word> =
+        [1, 2, 3, 4].iter().map(|&v| [Felt::new(v), ZERO, ZERO, ZERO]).collect();
+    let num_leaves = leaves.len();
+    let merkle_tree = MerkleTree::new(leaves).unwrap();
+    let merkle_root = merkle_tree.root();
+    let merkle_store: MerkleStore = MerkleStore::from(&merkle_tree);
+
+    let mut code = format!(
+        "
+                            use.std::collections::mmr
+                            use.miden::contracts::auth::basic->auth_tx
+                            use.miden::kernels::tx::prologue
+                            use.miden::kernels::tx::memory
+
+                            begin
+                                # leaf count -> mem[1000][0]
+                                push.{num_leaves} push.1000 mem_store
+
+                                # merkle root -> mem[1001]
+                                push.{} push.1001 mem_storew dropw
+                        ",
+        merkle_root.to_hex()
+    );
+
+    for pos in 0..(num_leaves as u64) {
+        let expected_element = merkle_store
+            .get_node(merkle_root, NodeIndex::new(2u8, pos).unwrap())
+            .unwrap()
+            .to_hex();
+        code += format!(
+            "
+            # get element at index `pos` from the merkle store in mem[1000] and push it to stack
+            push.1000 push.{pos} exec.mmr::get
+
+            # check the element matches what was inserted at `pos`
+            push.{expected_element} assert_eqw
+        "
+        )
+        .as_str();
+    }
+
+    code += "call.auth_tx::auth_tx_rpo_falcon512 end";
+
+    // Build the transaction
+    let program = ProgramAst::parse(&code).unwrap();
+    let tx_script = {
+        let account_auth = client.get_account_auth(regular_account.id()).unwrap();
+        let (pubkey_input, advice_map): (Word, Vec<Felt>) = match account_auth {
+            AuthSecretKey::RpoFalcon512(key) => (
+                key.public_key().into(),
+                key.to_bytes().iter().map(|a| Felt::new(*a as u64)).collect::<Vec<Felt>>(),
+            ),
+        };
+
+        let script_inputs = vec![(pubkey_input, advice_map)];
+        client.compile_tx_script(program, script_inputs, vec![]).unwrap()
+    };
+
+    let transaction_request = TransactionRequest::new(regular_account.id())
+        .with_authenticated_input_notes(note_args_map)
+        .with_tx_script(tx_script)
+        .extend_advice_map(advice_map)
+        .extend_merkle_store(merkle_store.inner_nodes());
 
     execute_tx_and_sync(&mut client, transaction_request).await;
 
@@ -201,7 +294,7 @@ async fn mint_custom_note(
     end
     "
     .replace("{recipient}", &recipient)
-    .replace("{note_type}", &Felt::new(NoteType::OffChain as u64).to_string())
+    .replace("{note_type}", &Felt::new(NoteType::Private as u64).to_string())
     .replace("{tag}", &Felt::new(note_tag.into()).to_string())
     .replace("{amount}", &Felt::new(10).to_string());
 
@@ -209,16 +302,9 @@ async fn mint_custom_note(
 
     let tx_script = client.compile_tx_script(program, vec![], vec![]).unwrap();
 
-    let transaction_request = TransactionRequest::new(
-        faucet_account_id,
-        vec![],
-        BTreeMap::new(),
-        vec![note.clone()],
-        vec![],
-        Some(tx_script),
-        None,
-    )
-    .unwrap();
+    let transaction_request = TransactionRequest::new(faucet_account_id)
+        .with_expected_output_notes(vec![note.clone()])
+        .with_tx_script(tx_script);
 
     let _ = execute_tx_and_sync(client, transaction_request).await;
     note
@@ -246,7 +332,7 @@ fn create_custom_note(
     let serial_num = rng.draw_word();
     let note_metadata = NoteMetadata::new(
         faucet_account_id,
-        NoteType::OffChain,
+        NoteType::Private,
         NoteTag::from_account_id(target_account_id, NoteExecutionHint::Local).unwrap(),
         Default::default(),
     )