+++ tests/integration/onchain_tests.rs
@@ -134,7 +134,7 @@ async fn test_onchain_accounts() {
     // First Mint necesary token
     println!("First client consuming note");
     let note =
-        mint_note(&mut client_1, target_account_id, faucet_account_id, NoteType::OffChain).await;
+        mint_note(&mut client_1, target_account_id, faucet_account_id, NoteType::Private).await;
 
     // Update the state in the other client and ensure the onchain faucet hash is consistent
     // between clients
@@ -151,7 +151,7 @@ async fn test_onchain_accounts() {
         &mut client_2,
         second_client_target_account_id,
         faucet_account_id,
-        NoteType::OffChain,
+        NoteType::Private,
     )
     .await;