+++ tests/integration/swap_transactions_tests.rs
@@ -129,7 +129,7 @@ async fn test_swap_fully_onchain() {
     let tx_request = client1.build_transaction_request(tx_template).unwrap();
 
     let expected_output_notes = tx_request.expected_output_notes().to_vec();
-    let expected_payback_note_details = tx_request.expected_partial_notes().to_vec();
+    let expected_payback_note_details = tx_request.expected_future_notes().to_vec();
     assert_eq!(expected_output_notes.len(), 1);
     assert_eq!(expected_payback_note_details.len(), 1);
 
@@ -333,13 +333,13 @@ async fn test_swap_offchain() {
             Asset::Fungible(offered_asset),
             Asset::Fungible(requested_asset),
         ),
-        NoteType::OffChain,
+        NoteType::Private,
     );
     println!("Running SWAP tx...");
     let tx_request = client1.build_transaction_request(tx_template).unwrap();
 
     let expected_output_notes = tx_request.expected_output_notes().to_vec();
-    let expected_payback_note_details = tx_request.expected_partial_notes().to_vec();
+    let expected_payback_note_details = tx_request.expected_future_notes().to_vec();
     assert_eq!(expected_output_notes.len(), 1);
     assert_eq!(expected_payback_note_details.len(), 1);
 
@@ -352,7 +352,7 @@ async fn test_swap_offchain() {
         .try_into()
         .unwrap();
     let tag =
-        build_swap_tag(NoteType::OffChain, offered_asset.faucet_id(), requested_asset.faucet_id());
+        build_swap_tag(NoteType::Private, offered_asset.faucet_id(), requested_asset.faucet_id());
     client2.add_note_tag(tag).unwrap();
     client2
         .import_note(NoteFile::NoteDetails(exported_note.into(), Some(tag)))